Logo images
These are the PNGs from the 1st img package that Ty handed over to Jeff.

& although this commit doesn't technically have much to do with it, I
believe we are each set up on our individual workstations at this point,
which is Awesome!  So I'm gonna take this opportunity to try out the
nifty trick Thomas showed us last class and say that this Fixes #13 !